+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-06-23
+### descriptions, comments, and converting draft items to issues
+- Added support for adding descriptions to items
+- Added support for adding comments to items
+- Added support for converting draft items to issues
+
 ## [0.3.4] - 2026-06-23
 ### Make binaries executable
 - Fixes an issue where the standalone binaries were not marked as executable on Linux and macOS